BAEL - 1060
Review comments incorporated
