+++ rxjava-2/src/test/java/com/baeldung/rxjava/RxJavaHooksUnitTest.java
@@ -235,67 +235,6 @@ public class RxJavaHooksUnitTest {
         assertTrue(hookCalled && initHookCalled);
     }
 
-    @Test
-    public void givenIOScheduler_whenCalled_shouldExecuteTheHooks() {
-
-        RxJavaPlugins.setInitIoSchedulerHandler((scheduler) -> {
-            initHookCalled = true;
-            return scheduler.call();
-        });
-
-        RxJavaPlugins.setIoSchedulerHandler((scheduler) -> {
-            hookCalled = true;
-            return scheduler;
-        });
-
-        Observable.range(1, 10)
-            .map(v -> v * 2)
-            .subscribeOn(Schedulers.io())
-            .test();
-        assertTrue(hookCalled && initHookCalled);
-    }
-
-    @Test
-    public void givenNewThreadScheduler_whenCalled_shouldExecuteTheHook() {
-
-        RxJavaPlugins.setInitNewThreadSchedulerHandler((scheduler) -> {
-            initHookCalled = true;
-            return scheduler.call();
-        });
-
-        RxJavaPlugins.setNewThreadSchedulerHandler((scheduler) -> {
-            hookCalled = true;
-            return scheduler;
-        });
-
-        Observable.range(1, 15)
-            .map(v -> v * 2)
-            .subscribeOn(Schedulers.newThread())
-            .test();
-        assertTrue(hookCalled && initHookCalled);
-    }
-
-    @Test
-    public void givenSingleScheduler_whenCalled_shouldExecuteTheHooks() {
-
-        RxJavaPlugins.setInitSingleSchedulerHandler((scheduler) -> {
-            initHookCalled = true;
-            return scheduler.call();
-        });
-
-        RxJavaPlugins.setSingleSchedulerHandler((scheduler) -> {
-            hookCalled = true;
-            return scheduler;
-        });
-
-        Observable.range(1, 10)
-            .map(v -> v * 2)
-            .subscribeOn(Schedulers.single())
-            .test();
-        assertTrue(hookCalled && initHookCalled);
-
-    }
-
     @After
     public void reset() {
         initHookCalled = false;
